+++ src/tyssue/behaviors/sheet/__init__.py
@@ -6,5 +6,12 @@ the completion of an event.
 """
 
 from .apoptosis_events import apoptosis  # noqa
-from .basic_events import division, face_elimination, type1_transition  # noqa
+from .basic_events import (  # noqa
+    IntersectionError,
+    check_intersections,
+    division,
+    face_elimination,
+    reconnect,
+    type1_transition,
+)
 from .delamination_events import constriction  # noqa