+++ src/tyssue/generation/__init__.py
@@ -7,8 +7,3 @@ from .hexagonal_grids import *  # noqa
 from .modifiers import *  # noqa
 from .shapes import *  # noqa
 from .utils import *  # noqa
-
-try:
-    from tyssue_cpp import mesh_generation  # noqa
-except ImportError:
-    print("C++ extensions are not available for this version")